Refactor mod for server-side compatibility
Separate client and server code to prevent crashes on dedicated
servers. Move client-specific classes to a new 'client' package and
update event registrations to avoid duplicate overlay issues.

Key changes:
- Create CommonConfig for shared configuration
- Implement ClientSetup for client-specific initialization
- Remove duplicate GUI overlay registrations
- Add @OnlyIn(Dist.CLIENT) annotations to client-only classes

This refactoring ensures the mod can be safely installed on servers
while maintaining full functionality when present on both client and
server.
