+++ src/main/java/com/jopgood/cfwinfo/items/JetpackInformation.java
@@ -1,5 +1,7 @@
 package com.jopgood.cfwinfo.items;
 
+import com.mojang.logging.LogUtils;
+import net.minecraft.client.Minecraft;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.player.Player;
@@ -8,10 +10,13 @@ import net.minecraft.world.item.ItemStack;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.jopgood.cfwinfo.data.JetpackDataManager;
+import com.jopgood.cfwinfo.common.data.JetpackDataManager;
+import org.slf4j.Logger;
 
 public class JetpackInformation {
-	    
+
+    private static final Logger LOGGER = LogUtils.getLogger();
+
     public static List<Component> updateTooltipWithSpacing(Player player) {
         List<Component> newTooltip = new ArrayList<>();
         
@@ -31,6 +36,17 @@ public class JetpackInformation {
         }
         
         return newTooltip;
+    }
+
+    // Method to update the jetpack information
+    private static void updateJetpackInformation() {
+        // Get the player entity (assuming it's available during client setup)
+        Minecraft mc = Minecraft.getInstance();
+        Player player = mc.player;
+        if (player != null) {
+            LOGGER.info("Player exists, executing JetpackInfo...");
+            JetpackInformation.updateTooltipWithSpacing(player);
+        }
     }
 }
